Refactor to use one api endpoint and one query parser
