+++ lib/collect/event.js
@@ -1,7 +1,7 @@
 const requestIp = require('request-ip')
 const anonymize = require('ip-anonymize')
 import {dataStoreEvent as datastore} from '../datastore'
-import {queryParserEvent as queryParser} from '../queryParser'
+import queryParser from '../queryParser'
 
 module.exports = async (req, res) => {
     const {
