+++ lib/collect/track.js
@@ -1,13 +1,13 @@
 const requestIp = require('request-ip')
 const anonymize = require('ip-anonymize')
-const {dataStoreTrack} = require('../datastore')
-const {queryParserTrack} = require('../queryParser')
+import {dataStoreTrack as datastore} from '../datastore'
+import queryParser from '../queryParser'
 
 module.exports = async (req, res) => {
     const { 
         programId, sessionId, locale, kind, affiliate, url,
         dimension1, dimension2, dimension3, dimension4, dimension5,
-    } = queryParserTrack(req.url)
+    } = queryParser(req.url)
 
     const now = new Date(Date.now()).toISOString();
 
@@ -54,7 +54,7 @@ module.exports = async (req, res) => {
 
     const tableName = kind
 
-    return dataStoreTrack
+    return datastore
         .insert(tableName, rows, now)
         .catch((err) => {
             console.error('ERROR:', err);
