+++ lib/datastore/track.js
@@ -21,9 +21,7 @@ const schema = {
     dimension3: 'string',
     dimension4: 'string',
     dimension5: 'string',
-    timestamp: 'datetime',
-    sessionId: 'string',
-    locale: 'string',
+    timestamp: 'datetime'
 }
 const schemaString = Object
     .entries(schema)
