--- lib/queryParser/index.js
@@ -1,2 +1,0 @@
-export { default as queryParserTrack } from "./track";
-export { default as queryParserEvent } from "./event";
