--- lib/queryParser/track.js
@@ -1,19 +1,0 @@
-const { parse } = require('url')
-
-export default url => {
-    const { query } = parse(url, true)
-
-    return {
-        programId: query.p,
-        sessionId: query.sid,
-        locale: query.locale,
-        kind: query.k,
-        affiliate: query.a,
-        url: query.url,
-        dimension1: query.d1,
-        dimension2: query.d2,
-        dimension3: query.d3,
-        dimension4: query.d4,
-        dimension5: query.d5,
-    }
-}
